1)Changed masterDb schema. it now supports precision and scale which are nullable columns for decimals.
2)Added alter table add column validation check for decimals and size assignment based on precision.

+++ src/Database/Constants.h
@@ -52,8 +52,8 @@ namespace Constants
 
     constexpr std::string_view WILDCARD = "*";
 
-    constexpr int INVALID_DECIMAL_PRECISION = -1;
-    constexpr int INVALID_DECIMAL_SCALE = -1;
+    constexpr int8_t INVALID_DECIMAL_PRECISION = -1;
+    constexpr int8_t INVALID_DECIMAL_SCALE = -1;
 
     constexpr int MAX_DECIMAL_PRECISION = 38;
     constexpr int MAX_DECIMAL_SCALE = 38;
